requirements.txt
fastapi
uvicorn
requests
pandas
pyyaml
apscheduler
jinja2
matplotlib
python-multipart
# aiofiles # potentially needed for fastapi static files
pyarrow
finance-datareader
yfinance